pyproject.toml
[build-system]
requires = ["hatchling>=1.17.1", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "coffea"
description = "Basic tools and wrappers for enabling not-too-alien syntax when running columnar Collider HEP analysis."
readme = "README.rst"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.8"
authors = [
  { name = "Lindsey Gray", email = "lagray@fnal.gov" },
  { name = "Nick Smith", email = "ncsmith@fnal.gov" },
]
maintainers = [
  { name = "Lindsey Gray", email = "lagray@fnal.gov" },
  { name = "Nick Smith", email = "ncsmith@fnal.gov" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: BSD License",
  "Operating System :: MacOS",
  "Operating System :: POSIX",
  "Operating System :: Unix",
  "Operating System :: Microsoft :: Windows",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Information Analysis",
  "Topic :: Scientific/Engineering :: Mathematics",
  "Topic :: Scientific/Engineering :: Physics",
  "Topic :: Software Development",
  "Topic :: Utilities",
]
dependencies = [
  "awkward>=2.6.7",
  "uproot>=5.3.11",
  "dask[array]>=2024.3.0;python_version>'3.8'",
  "dask[array]>=2023.4.0;python_version<'3.9'",
  "dask-awkward>=2024.9.0",
  "dask-histogram>=2024.9.1",
  "vector>=1.4.1",
  "correctionlib>=2.6.0",
  "pyarrow>=6.0.0",
  "fsspec-xrootd>=0.2.3",
  "matplotlib>=3",
  "numba>=0.58.1",
  "numpy>=1.22",
  "scipy>=1.1.0",
  "tqdm>=4.27.0",
  "lz4",
  "cloudpickle>=1.2.3",
  "toml>=0.10.2",
  "mplhep>=0.1.18",
  "packaging",
  "pandas",
  "hist>=2",
  "cachetools",
  "requests",
  "aiohttp",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/coffeateam/coffea"
"Bug Tracker" = "https://github.com/coffeateam/coffea/issues"

[project.optional-dependencies]
dask = [
  "distributed>=2024.3.0;python_version>'3.8'",
  "distributed>=2023.4.0;python_version<'3.9'",
  "bokeh!=3.0.*,>=2.4.2",
]
spark = [
  "ipywidgets",
  "pyspark>=3.3.0",
  "jinja2",
]
parsl = [
  "parsl>=2022.12.1"
]
servicex = [
  "aiostream",
  "tenacity",
  "servicex>=2.5.3",
  "func-adl_servicex",
]
rucio = [
  "rucio-clients>=32;python_version>'3.8'",
  "rucio-clients<32;python_version<'3.9'",
]
dev = [
  "pre-commit",
  "flake8",
  "black",
  "pytest",
  "pytest-cov",
  "pytest-mpl",
  "pytest-asyncio",
  "pytest-mock",
  "sphinx<8",
  "nbsphinx",
  "sphinx-rtd-theme",
  "sphinx-automodapi",
  "sphinx-copybutton>=0.3.2",
  "pyinstrument",
  "ipython",
  "distributed>=2023.4.0",
]

#[project.entry-points."dask.sizeof"]
#coffea = "coffea.sizeof:register"

[tool.hatch]
version.source = "vcs"
build.hooks.vcs.version-file = "src/coffea/_version.py"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
addopts = [
  "-rfesxX",
  "-v",
]
log_cli_level = "DEBUG"
filterwarnings = [
  "ignore:There is no current event loop",
]

[tool.ruff]
line-length = 160

[tool.ruff.lint]
ignore = ["F403", "F405", "E402"]
